Put compile output into build root directories
We were putting build output into CMAKE_CURRENT_BINARY_DIR/{bin,lib}
which makes it relative to the build script that command is in. Instead
make it relative to CMAKE_BINARY_DIR which places it relative to the
ultimate root build directory. The two paths differ when using this
project via add_subdirectory() from a parent project, for example.