Relicense source to GPL 3 or later
I corresponded with Pierre L'Ecuyer to confirm that the copyright page
on the TestU01 website [1] is intended to supercede the non-commercial
license found in the source distribution. The new license is:

    GPL-3.0-or-later AND LicenseRef-testu01

The copyright page is included as LicenseRef-testu01 because it has
additional terms as permitted under GPL 3 section 7 ('Additional
Terms').

I also fixed some missing license information for files which were taken
from an external source and have their own license (which is either a
public domain statement or an ad-hoc permissive license).

The in-source changes are only in the C source files. Everything else is
documented in .reuse/dep5. I plan to make changes to header
file/documentation generation so didn't change the TeX sources yet.

[1] http://simul.iro.umontreal.ca/testu01/copyright.html